# openapi/dist/openapi.yaml
openapi: 3.1.0
info:
  version: 1.0.0
  title: Reel Fake Api
  description: A REST api for buying or renting a movie dvd (only for educational purposes)
  license:
    name: MIT License
    url: https://mit-license.org
  contact:
    name: Pratap Reddy
    email: pratapreddy15@gmail.com
servers:
  - url: http://reelfake.cloud/api
    description: Production server
  - url: http://127.0.0.1:8000/api
    description: Local server
security:
  - cookieAuth: []
paths:
...
  /stores:
    get:
      summary: Get list of stores
      operationId: getStores
      tags:
        - store
      responses:
        '200':
          description: Successfully returned an array of storeItem
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/storeItem'
                  length:
                    type: integer
        '400':
          $ref: '#/components/responses/badRequest'
        '500':
          $ref: '#/components/responses/internalServerError'
    post:
      summary: Create store
      operationId: createStore
      tags:
        - store
      requestBody:
        description: Payload for the new store
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  properties:
                    storeManagerId:
                      type: integer
                      format: number
                    phoneNumber:
                      type: string
                    address:
                      $ref: '#/components/schemas/address'
                - type: object
                  properties:
                    phoneNumber:
                      type: string
                    address:
                      $ref: '#/components/schemas/address'
                    storeManager:
                      type: object
                      properties:
                        firstName:
                          type: string
                        lastName:
                          type: string
                        email:
                          type: string
                          format: email
                        phoneNumber:
                          type: string
                        avatar:
                          type: string
                        address:
                          $ref: '#/components/schemas/address'
      responses:
        '201':
          description: Successfully created store
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: number
                  storeManager:
                    type: object
                    properties:
                      id:
                        type: integer
                        format: number
                      firstName:
                        type: string
                      lastName:
                        type: string
                      email:
                        type: string
                        format: email
                      active:
                        type: boolean
                      avatar:
                        type: string
                      address:
                        $ref: '#/components/schemas/address'
                  phoneNumber:
                    type: string
                  address:
                    $ref: '#/components/schemas/address'
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '500':
          $ref: '#/components/responses/internalServerError'
  /stores/{id}:
    get:
      summary: Get store detail using the id
      operationId: getStoreById
      tags:
        - store
      parameters:
        - in: path
          name: id
          required: true
          description: The id of the store
          schema:
            type: integer
        - in: query
          name: includeStaff
          required: false
          description: |
            Whether or not to include the staff employed at the store. Accepted values are true, false, yes, no, 0 and 1.
            If includeStaff value is truthy then the store object will contain staff array.
            If this is not present in query then it will default to false. If any value other than the accepted ones are
            given, then 400 is returned.
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Successfully returned the store data with or without the staff
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/storeDetail'
        '400':
          $ref: '#/components/responses/badRequest'
        '404':
          $ref: '#/components/responses/notFound'
        '500':
          $ref: '#/components/responses/internalServerError'
    put:
      summary: Update store
      operationId: updateStore
      tags:
        - store
      parameters:
        - in: path
          name: id
          required: true
          description: Id of the store
          schema:
            type: integer
      requestBody:
        description: Payload for the new store
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                storeManagerId:
                  type: integer
                  format: number
                phoneNumber:
                  type: string
                address:
                  $ref: '#/components/schemas/address'
      responses:
        '204':
          description: Successfully updated the store
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '500':
          $ref: '#/components/responses/internalServerError'
    delete:
      summary: Delete the store
      operationId: deleteStore
      tags:
        - store
      parameters:
        - in: path
          name: id
          required: true
          description: The id of the store
          schema:
            type: integer
        - in: query
          name: forceDelete
          required: false
          description: |
            A boolean flag to indicate whether or not to delete the staff associated with the store.
            Default is false.
          schema:
            type: boolean
            default: false
      responses:
        '204':
          description: Successfully deleted the store
        '400':
          $ref: '#/components/responses/badRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/notFound'
        '500':
          $ref: '#/components/responses/internalServerError'
  /stores/{id}/stock:
    get:
      summary: Get the number of in-stock movies in a given store
      operationId: getStockCount
      tags:
        - store
      parameters:
        - in: path
          name: id
          required: true
          description: The id of the store
          schema:
            type: integer
      responses:
        '200':
          description: Successfully returned the number of in-stock movies in a given store
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  stockCount:
                    type: string
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/unauthorized'
  /stores/{id}/movies:
    get:
      summary: Get movies in a store
      operationId: getMoviesInStore
      tags:
        - store
      parameters:
        - in: path
          name: id
          required: true
          description: Store id
          schema:
            type: integer
        - in: query
          name: page
          required: false
          description: |
            The page number (which starts with 1) to fetch the set of in-stock movies from.
            The movies are split into multiple pages with each page having a limited number of items.
            This query also accepts first and last.
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: Successfully returned movies in a store
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          format: number
                        stock:
                          type: integer
                          format: number
                        movie:
                          $ref: '#/components/schemas/movieSummary'
                  length:
                    type: integer
                  pagination:
                    $ref: '#/components/schemas/pagination'
        '400':
          $ref: '#/components/responses/badRequest'
        '404':
          $ref: '#/components/responses/notFound'
        '500':
          $ref: '#/components/responses/internalServerError'
  /stores/{id}/staff:
    get:
      summary: Get staff employed in a store
      operationId: getStaffInStore
      tags:
        - store
      parameters:
        - in: path
          name: id
          required: true
          description: The id of the store
          schema:
            type: integer
      responses:
        '200':
          description: Successfully returned staff employed in a store
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          format: number
                        firstName:
                          type: string
                        lastName:
                          type: string
                        email:
                          type: string
                        addressId:
                          type: integer
                          format: number
                        storeId:
                          type: integer
                          format: number
                        active: true
                        phoneNumber:
                          type: string
                        avatar:
                          type: string
                        isStoreManager:
                          type: boolean
                        store:
                          type: object
                          properties:
                            id:
                              type: integer
                              format: number
                            storeManagerId:
                              type: integer
                              format: number
                            addressId:
                              type: integer
                              format: number
                            phoneNumber:
                              type: string
                  length:
                    type: integer
        '400':
          $ref: '#/components/responses/badRequest'
        '404':
          $ref: '#/components/responses/notFound'
        '500':
          $ref: '#/components/responses/internalServerError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      name: auth_token
      in: cookie
  schemas:
...
    pagination:
      type: object
      properties:
        pageNumber:
          type: integer
          description: Current page number
        totalPages:
          type: integer
          description: Total pages available for pagination
        totalItems:
          type: integer
          description: Total items across all the pages
        itemsPerPage:
          type: integer
          description: Maximum number of items that the page can have
        next:
          type: string
          description: Link to the next page
        prev:
          type: string
          description: Link to the previous page
        first:
          type: string
          description: Link to the first page
        last:
          type: string
          description: Link to the last page
...
    address:
      type: object
      properties:
        id:
          type: integer
          format: number
          readOnly: true
        addressLine:
          type: string
        cityName:
          type: string
        stateName:
          type: string
        country:
          type: string
        postalCode:
          type: string
...
    movieSummary:
      type: object
      properties:
        id:
          type: integer
          description: Unique id of the resource
        title:
          type: string
          description: The title of the movie
        runtime:
          type: integer
          description: The length of the movie
        releaseDate:
          type: string
          description: Date on which the movie is or will be released
          format: date
        genres:
          type: array
          description: |
            The genres that the movie belongs to. The valid genres are action, adventure, animation, 
            comedy, crime, documentary, drama, family, fantasy, history, horror, music , mystery, romance, 
            science fiction, thriller, war and western.
          items:
            type: string
        countriesOfOrigin:
          type: array
          description: The countries of origin of the movie
          items:
            type: string
        language:
          type: string
          description: The language that the movie is in
        popularity:
          type: number
          description: A number indicating how popular the movie is
          format: float
        ratingAverage:
          type: number
          description: The average rating of the movie
          format: float
        ratingCount:
          type: integer
          description: The number of ratings that the movie has got
        posterUrl:
          type: string
          description: Link to the movie poster
          format: uri
        rentalRate:
          type: number
          description: The rental rate of the movie
          format: float
...
    storeItem:
      type: object
      properties:
        id:
          type: integer
        addressLine:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
        phoneNumber:
          type: string
    storeDetail:
      type: object
      properties:
        id:
          type: integer
          format: number
        phoneNumber:
          type: string
        storeManager:
          type: object
          properties:
            id:
              type: integer
              format: number
            firstName:
              type: string
            lastName:
              type: string
            email:
              type: string
            phoneNumber:
              type: string
            storeId:
              type: integer
              format: number
            active:
              type: boolean
            avatar:
              type: string
            address:
              $ref: '#/components/schemas/address'
        address:
          $ref: '#/components/schemas/address'
...
  responses:
    badRequest:
      description: The request does not have the data that api needed
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    internalServerError:
      description: Server encountered unexpeted error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    unauthorized:
      description: Api key is missing or invalid
      headers:
        WWW_Authenticate:
          schema:
            type: string
    notFound:
      description: Requested resource does not exist
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    forbidden:
      description: User do not have permission to perform the operation
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string